Add files via upload
This is a python code for operating forward step-wise regression analysis. In forward step-wise regression, a regression analysis is done for each individual input parameter and find out the input parameter contains lowest p value. Then the input variable is added to the model. After this, another regression analysis is carried out using the selected parameter from the previous step with one of each of the remaining input variable one at a time and same as before, find the latest added input variable for which p value is minimum. The input variable is then added to the model as 2nd input variable. Thus, the analysis goes on until minimum p value found for latest added input variable is greater than the confidence level (0.05 mostly). This code is written to perform this exact step for any number of input variables.